Added the MVCExample program. I suggest using it to get a handle on how databinding in java swing works and how the mvc structure is used. The View has a controller which has a model. A control in the View is bound to a field in the model of the controller and will be updated whenever the model is updated and visa versa. Any requests that the View has that would require anything beyond this basic link is handled through the controller. This is seen in the chester button. The View has a clicked event that passes on to the controller what it wants it to do. The controller then tells the model what to do and the model performs the requested action and updates itself, resulting in updated information being displayed on the view.
I attempted to model this as closely as i could the actual implementation we currently have in the pizza system, minus the web service and database accesses. This means that there are 2 interfaces. One for the controller and one for the model. There are two abstract classes which implement the interface and any methods that are common between the children. In both cases this is the property changed code. The models then extend the Abstract classes and simply supply the property information and any methods specific to the implementation of the model. Any method that is not present in the interface and are not public facing properties should be private.

A property is simply a class variable that has both a getter and a setter method.